+++ template_variables/src/epilot_template_variables/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.3.0"
-    gen_version: str = "2.390.0"
-    user_agent: str = "speakeasy-sdk/python 1.3.0 2.390.0 1.0.0 epilot-template-variables"
+    sdk_version: str = "1.3.1"
+    gen_version: str = "2.390.6"
+    user_agent: str = "speakeasy-sdk/python 1.3.1 2.390.6 1.0.0 epilot-template-variables"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None