+++ template_variables/RELEASES.md
@@ -43,3 +43,11 @@ Based on:
 - Speakeasy CLI 1.357.3 (2.390.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.3.0] template_variables
+
+## 2024-08-09 00:25:11
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.358.0 (2.390.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.3.1] template_variables